@@ -26,13 +26,18 @@ dependencies:
   google_generative_ai: ^0.4.6
   loading_animation_widget: ^1.3.0
   table_calendar: ^3.1.3
-  camera: ^0.11.1
+  camera: ^0.10.5+9
   permission_handler: ^11.4.0
   firebase_analytics: ^11.4.5
   firebase_core: ^3.13.0
   cloud_firestore: ^5.6.6
   http: ^1.3.0
-  google_maps_flutter: ^2.5.3
+  flutter_dotenv: ^5.1.0
+  fl_chart: ^0.69.0
+  lottie: ^3.1.2
+  flutter_animate: ^4.5.0
+  flutter_staggered_animations: ^1.1.1
+  google_maps_flutter: ^2.10.0
 
 dev_dependencies:
   flutter_test:
@@ -44,8 +49,10 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
+    - .env
     - assets/images/
     - assets/transactions.json
+    - assets/mock_transactions.json
     - assets/images/shirts/
     - assets/images/cats/
     - assets/images/hats/
